Preformace update 1
Used the inspector to reduce script runtimes by 17x, removed many unnesseary array conversion and variable declarations.

Forces now check in their gameObjects to a active forces list in Beam, allowing for a full removal of the serch function.

GC is till kind of high due to particle memory alloction, so that is next on the list